Created vector2 double and replaced it in ship and ship control

+++ src/cx/it/hyperbadger/spacezombies/explorer/Ship.java
@@ -11,18 +11,18 @@ import java.util.ArrayList;
 
 import org.lwjgl.opengl.Display;
 import org.lwjgl.opengl.GL11;
-import org.lwjgl.util.vector.Vector2f;
 import org.newdawn.slick.opengl.Texture;
 import org.newdawn.slick.opengl.TextureLoader;
 
 import cx.it.hyperbadger.spacezombies.Game;
+import cx.it.hyperbadger.spacezombies.Vector2d;
 
 public class Ship extends Mass implements Drawable{
 	private Texture texture = null;
 	private float rotation = 0;
 	private boolean first = true;
 	private int h,w;
-	private Vector2f velocity = new Vector2f();
+	private Vector2d velocity = new Vector2d();
 	public Ship(int mass, int x, int y, String texture){
 		super(mass,x,y,"Ship");
 		try {
@@ -55,15 +55,15 @@ public class Ship extends Mass implements Drawable{
 		GL11.glRotatef(-rotation, 0f, 0f, 1f);
 		GL11.glTranslatef(-(float)Display.getWidth()/2-w/2, -(float)Display.getHeight()/2-h/2, 0);
 	}
-	public void applyForce(Vector2f force){
+	public void applyForce(Vector2d force){
 		velocity.x = velocity.x + (force.x*Game.delta/1000)/this.mass;
 		velocity.y = velocity.y + (force.y*Game.delta/1000)/this.mass;
 	}
 
 	public void move(ArrayList<Mass> attracts){
-		Vector2f planetForce = new Vector2f(0,0);
+		Vector2d planetForce = new Vector2d(0,0);
 		for(Mass m: attracts){
-			Vector2f.add(planetForce, getVectorForce(this,m), planetForce);
+			Vector2d.add(planetForce, getVectorForce(this,m), planetForce);
 		}
 		//we dont apply force initially, shit fucks up
 		if(!first){
@@ -73,32 +73,32 @@ public class Ship extends Mass implements Drawable{
 		this.x = this.x + velocity.x*Game.delta/1000;
 		this.y = this.y + velocity.y*Game.delta/1000;
 	}
-	private Vector2f getVectorForce(Mass from, Mass to){
+	private Vector2d getVectorForce(Mass from, Mass to){
 		//from ship
 		//to planet
-		Vector2f ship = new Vector2f();
+		Vector2d ship = new Vector2d();
 		ship.x = (float)this.getX();
 		ship.y = (float)this.getY();
-		Vector2f mass = new Vector2f();
+		Vector2d mass = new Vector2d();
 		mass.x = (float)to.getX();
 		mass.y = (float)to.getY();
-		Vector2f force = new Vector2f();
-		Vector2f.sub(mass, ship, force);
-		Vector2f unitVector = new Vector2f();
+		Vector2d force = new Vector2d();
+		Vector2d.sub(mass, ship, force);
+		Vector2d unitVector = new Vector2d();
 		float divisor = (float) Math.sqrt(Math.pow(force.x, 2)+Math.pow(force.y, 2));
 		unitVector.set(force.x/divisor,force.y/divisor);
 		float magnitude = (float) this.getForceMagnitude(to);
-		Vector2f magVec = new Vector2f();
+		Vector2d magVec = new Vector2d();
 		magVec.set(unitVector.x*magnitude,unitVector.y*magnitude);
 		return magVec;
 	}
 	private void calculateAngle(){
-		Vector2f velocityUnit = new Vector2f();
-		Vector2f up = new Vector2f(0,1);
+		Vector2d velocityUnit = new Vector2d();
+		Vector2d up = new Vector2d(0,1);
 		float velocityMag = (float) Math.sqrt(Math.pow(velocity.x, 2)+Math.pow(velocity.y, 2));
 		if(velocityMag!=0){
 			velocityUnit.set(velocity.x/velocityMag,velocity.y/velocityMag);
-			rotation = (float) Math.toDegrees(Vector2f.angle(up, velocityUnit))-90;
+			rotation = (float) Math.toDegrees(Vector2d.angle(up, velocityUnit))-90;
 		}
 	}
 	public float getVelocity(){
